+++ CHANGELOG.md
@@ -2,16 +2,26 @@
 
 This file documents any relevant changes done to ViUR-core since version 3.0.0.
 
-
-## [develop] - Current development version
+## [main] - Current main version
 
 ### Added
 
 ### Changed
 
 ### Fixed
+- required=True bones could still be set to empty if omitted from the request
+- treeNodeBone enforcing "_rootNode" suffix on it's kind which this isn't true for ViUR3 anymore.
 
 ### Removed
+
+## [3.1.4] - Current development version
+
+### Fixed
+- Fix unused language parameter of `utils.seoUrlToEntry()` (#439)
+- Subdependencies updated (#442)
+
+### Removed
+- Remove class `errors.ReadFromClientError`. Replaced by the new dataclass `bones.bone.ReadFromClienError`.  (#437, #443)
 
 ## [3.1.3]